+++ doc/changes.rst
@@ -5,6 +5,12 @@ desitarget Change Log
 1.0.2 (unreleased)
 ------------------
 
+* Refactor data model and I/O for the QSO zcats [`PR #737`_]. Includes:
+    * New directory structures.
+    * Allow filenames to be passed instead of TILE/NIGHT combinations.
+    * Always run QuasarNP (QN) by default.
+    * Update which redrock columns are populated in the output files.
+    * Only include output columns for afterburners that are run.
 * Add option to use Gaia EDR3 when selecting GFAs [`PR #734`_]. Also:
     * Initialize ``SUBPRIORITY`` with better random seeds.
          * This is crucial when parallelizing across HEALPixels.
@@ -17,6 +23,7 @@ desitarget Change Log
 
 .. _`PR #733`: https://github.com/desihub/desitarget/pull/733
 .. _`PR #734`: https://github.com/desihub/desitarget/pull/734
+.. _`PR #737`: https://github.com/desihub/desitarget/pull/737
 
 1.0.1 (2021-05-14)
 ------------------